+++ Formatting/Pixoo.Weather.format.ps1
@@ -1,0 +1,1 @@
+Write-FormatView -TypeName Pixoo.Weather -Property Weather, CurTemp, MaxTemp, MinTemp, Pressure, WindSpeed, Visibility -AutoSize